adding links

+++ inc/composents/C_conter.hpp
@@ -15,7 +15,7 @@ class C_conter: public Chipset
 public:
 	C_conter();
 private:
-	nts::Tristate calculate(int index);
+	nts::Tristate calculate(size_t index);
 };
 
 #endif /* !C_CONTER_HPP_ */
