+++ inc/composents/C_invert.hpp
@@ -15,7 +15,7 @@ class C_invert: public Chipset
 public:
 	C_invert();
 private:
-	nts::Tristate calculate(int index);
+	nts::Tristate calculate(size_t index);
 };
 
 #endif /* !C_INVERT_HPP_ */
